requirements.txt
django
gunicorn
psycopg2-binary
whitenoise
dj-database-url
python-dotenv